May 7th 2020 Update
Main changes that were made since the last update included changing the Customers_Also_Bought to a classical component rather than a functional because I needed to store state. In addition I managed to render the lower box on the Customers_Also_Bought component and added some CSS  to it and I stored all my items in an object in state rather than hard coding them in as I did before.